requirements.txt
Flask==2.0.1
Keras==2.4.3
numpy==1.21.1
tensorflow-gpu==2.5.0
